Corrected radians -> degrees; added directory for pure julia: julia_version/
